+++ orchagent/intfsorch.cpp
@@ -1,14 +1,13 @@
-#include "intfsorch.h"
-
-#include "ipprefix.h"
-#include "logger.h"
-
-#include "assert.h"
+#include <assert.h>
 #include <fstream>
 #include <sstream>
 #include <map>
-
 #include <net/if.h>
+
+#include "intfsorch.h"
+#include "ipprefix.h"
+#include "logger.h"
+#include "sainet.h"
 
 extern sai_object_id_t gVirtualRouterId;
